Added security measures to prevent spam of course creation and unauthorized deletion of data

+++ lib/actions.ts
@@ -36,6 +36,13 @@ export async function createRating(previousState: any, formData: FormData) {
 export async function deleteRating(previousState: any, rating: Rating) {
   if (!isLoggedIn()) return "You must be logged in to delete a rating";
 
+  // Check if the user is the author of the rating
+  const session = await auth();
+  const authorId = session?.user?.id;
+  if (authorId !== rating.authorId) {
+    return "You can only delete your own ratings";
+  }
+
   try {
     await prisma.rating.delete({
       where: { id: rating.id },
@@ -50,34 +57,78 @@ export async function getSession() {
   const session = await auth();
   return session;
 }
-
+/**
+ * Adds a course to the database
+ * 
+ * @param schoolId id of the school
+ * @param code The course code
+ * @returns an optional error message and the course id
+ */
 export async function addCourse(schoolId: string, code: string) {
-  if (!isLoggedIn()) return JSON.stringify({ error: "You must be logged in to add a course", courseId: null });
+  // Check if user is logged in
+  const session = await auth();
+  if (!session || !session.user) {
+    return JSON.stringify({
+      error: "You must be logged in to add a course",
+      courseId: null,
+    });
+  }
 
+  const authorId = session.user.id;
+  if (!authorId) {
+    return JSON.stringify({
+      error: "Invalid user",
+      courseId: null,
+    });
+  }
+
+  // Validate input
   if (code.trim() === "" || schoolId.trim() === "") {
     return JSON.stringify({
       error: "Missing required fileds",
       courseId: null,
     });
   }
+
+  // Check if user has added 3 courses in the last 24 hours
+  const coursesAddedLastDay = await prisma.course.count({
+    where: {
+      authorId,
+      createdAt: {
+        gte: new Date(Date.now() - 24 * 60 * 60 * 1000), // last 24 hours
+      }
+    }
+  });
+
+  if (coursesAddedLastDay >= 3) {
+    return JSON.stringify({
+      error: "You can only add 3 courses in a 24 hour period.",
+      courseId: null,
+    });
+  }
+  // Check for duplicates
   const duplicates = await prisma.course.findFirst({
     where: {
       code,
       schoolId,
     },
   });
+
   if (duplicates) {
     return JSON.stringify({
       error: "Course already exists",
       courseId: null,
     });
   }
+
+  // Create the course
   let result;
   try {
     result = await prisma.course.create({
       data: {
         code,
         schoolId,
+        authorId,
       },
     });
   } catch (error) {
@@ -86,6 +137,7 @@ export async function addCourse(schoolId: string, code: string) {
       courseId: null,
     });
   }
+  // Revalidate the path to show the new course immediately and return the course id
   revalidatePath("/addCourse");
   return JSON.stringify({
     error: null,
